Remove todo
Turns out, what we want (for changelog diffs etc) is not the
previously deployed version in this environment (staging), but rather
the currently active version in production, because the difference
between that and the now newly deployed staging version is what
actually is changed by a staging deployment.

And that's not something we can determine inside the staging cluster,
but must be passed in from the deployment machinery.